mask type schema
