+++ crates/grida-canvas/examples/golden_type_stroke.rs
@@ -29,7 +29,7 @@ async fn scene() -> Scene {
         stroke_align: StrokeAlign::Outside,
         opacity: 1.0,
         blend_mode: LayerBlendMode::default(),
-        mask_type: LayerMaskType::default(),
+        mask: None,
         effects: LayerEffects::default(),
     };
     let text_outside_id = text_outside.id.clone();
@@ -55,7 +55,7 @@ async fn scene() -> Scene {
         stroke_align: StrokeAlign::Center,
         opacity: 1.0,
         blend_mode: LayerBlendMode::default(),
-        mask_type: LayerMaskType::default(),
+        mask: None,
         effects: LayerEffects::default(),
     };
     let text_center_id = text_center.id.clone();
@@ -81,7 +81,7 @@ async fn scene() -> Scene {
         stroke_align: StrokeAlign::Inside,
         opacity: 1.0,
         blend_mode: LayerBlendMode::default(),
-        mask_type: LayerMaskType::default(),
+        mask: None,
         effects: LayerEffects::default(),
     };
     let text_inside_id = text_inside.id.clone();
