+++ crates/grida-canvas/examples/grida_booleans.rs
@@ -60,7 +60,7 @@ async fn demo_booleans() -> Scene {
             active: true,
             opacity: 1.0,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: Some(AffineTransform::new(start_x + spacing * 2.0, y_offset, 0.0)),
             op: BooleanPathOperation::Union,
@@ -124,7 +124,7 @@ async fn demo_booleans() -> Scene {
             active: true,
             opacity: 1.0,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: Some(AffineTransform::new(start_x + spacing * 2.0, y_offset, 0.0)),
             op: BooleanPathOperation::Intersection,
@@ -188,7 +188,7 @@ async fn demo_booleans() -> Scene {
             active: true,
             opacity: 1.0,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: Some(AffineTransform::new(start_x + spacing * 2.0, y_offset, 0.0)),
             op: BooleanPathOperation::Difference,
@@ -252,7 +252,7 @@ async fn demo_booleans() -> Scene {
             active: true,
             opacity: 1.0,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: Some(AffineTransform::new(start_x + spacing * 2.0, y_offset, 0.0)),
             op: BooleanPathOperation::Xor,
