+++ crates/grida-canvas/examples/grida_vector.rs
@@ -31,7 +31,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(start_x, base_y, 0.0),
                 fills: Paints::default(),
@@ -62,7 +62,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(start_x + spacing * 1.0, base_y, 0.0),
                 fills: Paints::default(),
@@ -95,7 +95,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(start_x + spacing * 2.0, base_y, 0.0),
                 fills: Paints::default(),
@@ -127,7 +127,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(start_x + spacing * 3.0, base_y, 0.0),
                 fills: Paints::default(),
@@ -160,7 +160,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(start_x + spacing * 4.0, base_y, 0.0),
                 fills: Paints::default(),
@@ -198,7 +198,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(
                     start_x + spacing * 0.0,
@@ -236,7 +236,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(
                     start_x + spacing * 1.0,
@@ -272,7 +272,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(
                     start_x + spacing * 0.0,
@@ -309,7 +309,7 @@ async fn demo_vectors() -> Scene {
                 active: true,
                 opacity: 1.0,
                 blend_mode: LayerBlendMode::default(),
-                mask_type: LayerMaskType::default(),
+                mask: None,
                 effects: LayerEffects::default(),
                 transform: AffineTransform::new(
                     start_x + spacing * 1.0,
