+++ crates/grida-canvas/src/io/io_figma.rs
@@ -580,7 +580,7 @@ impl FigmaConverter {
             active: component.visible.unwrap_or(true),
             opacity: Self::convert_opacity(component.visible),
             blend_mode: Self::convert_blend_mode(component.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform,
             size,
             corner_radius: Self::convert_corner_radius(
@@ -669,7 +669,7 @@ impl FigmaConverter {
             active: instance.visible.unwrap_or(true),
             opacity: Self::convert_opacity(instance.visible),
             blend_mode: Self::convert_blend_mode(instance.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform,
             size,
             corner_radius: Self::convert_corner_radius(
@@ -710,7 +710,7 @@ impl FigmaConverter {
             active: section.visible.unwrap_or(true),
             opacity: Self::convert_opacity(section.visible),
             blend_mode: LayerBlendMode::PassThrough,
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform: Self::convert_transform(section.relative_transform.as_ref()),
             size: Self::convert_size(section.size.as_ref()),
             corner_radius: RectangularCornerRadius::zero(),
@@ -823,7 +823,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform,
             size,
             corner_radius: Self::convert_corner_radius(
@@ -975,7 +975,7 @@ impl FigmaConverter {
             stroke_align: StrokeAlign::Inside,
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: Self::convert_effects(&origin.effects),
         }))
     }
@@ -993,7 +993,7 @@ impl FigmaConverter {
                     active: origin.visible.unwrap_or(true),
                     opacity: Self::convert_opacity(origin.visible),
                     blend_mode: Self::convert_blend_mode(origin.blend_mode),
-                    mask_type: LayerMaskType::default(),
+                    mask: None,
                     effects: Self::convert_effects(&origin.effects),
                     transform: AffineTransform::identity(),
                     fills: self.convert_fills(Some(&origin.fills)),
@@ -1018,7 +1018,7 @@ impl FigmaConverter {
                     active: origin.visible.unwrap_or(true),
                     opacity: Self::convert_opacity(origin.visible),
                     blend_mode: Self::convert_blend_mode(origin.blend_mode),
-                    mask_type: LayerMaskType::default(),
+                    mask: None,
                     effects: Self::convert_effects(&origin.effects),
                     transform: AffineTransform::identity(),
                     fills: Paints::default(),
@@ -1040,7 +1040,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform: Self::convert_transform(origin.relative_transform.as_ref()),
             size: Self::convert_size(origin.size.as_ref()),
             corner_radius: RectangularCornerRadius::zero(),
@@ -1088,7 +1088,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: Self::convert_effects(&origin.effects),
             transform: Some(transform),
             op: op,
@@ -1122,7 +1122,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: Self::convert_effects(&origin.effects),
             transform,
             size,
@@ -1158,7 +1158,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: Self::convert_effects(&origin.effects),
             transform,
             size,
@@ -1196,7 +1196,7 @@ impl FigmaConverter {
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
             effects: Self::convert_effects(&origin.effects),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform,
             size,
             fills: self.convert_fills(Some(&origin.fills)),
@@ -1236,7 +1236,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: Self::convert_effects(&origin.effects),
             transform,
             size,
@@ -1270,7 +1270,7 @@ impl FigmaConverter {
             active: origin.visible.unwrap_or(true),
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: Self::convert_effects(&origin.effects),
             transform,
             size,
@@ -1311,7 +1311,7 @@ impl FigmaConverter {
             children,
             opacity: Self::convert_opacity(origin.visible),
             blend_mode: Self::convert_blend_mode(origin.blend_mode),
-            mask_type: LayerMaskType::default(),
+            mask: None,
         }))
     }
 }
