+++ crates/grida-canvas/src/node/factory.rs
@@ -58,7 +58,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform: AffineTransform::identity(),
             size: Self::DEFAULT_SIZE,
             corner_radius: RectangularCornerRadius::zero(),
@@ -79,8 +79,8 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
+            mask: None,
             effects: LayerEffects::default(),
-            mask_type: LayerMaskType::default(),
             transform: AffineTransform::identity(),
             size: Self::DEFAULT_SIZE,
             inner_radius: None,
@@ -103,7 +103,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: AffineTransform::identity(),
             size: Size {
@@ -123,6 +123,10 @@ impl NodeFactory {
             id: self.id(),
             name: None,
             active: true,
+            opacity: Self::DEFAULT_OPACITY,
+            blend_mode: LayerBlendMode::default(),
+            mask: None,
+            effects: LayerEffects::default(),
             transform: AffineTransform::identity(),
             width: None,
             height: None,
@@ -136,10 +140,6 @@ impl NodeFactory {
             strokes: Paints::default(),
             stroke_width: 0.0,
             stroke_align: Self::DEFAULT_STROKE_ALIGN,
-            opacity: Self::DEFAULT_OPACITY,
-            blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
-            effects: LayerEffects::default(),
         }
     }
 
@@ -153,7 +153,7 @@ impl NodeFactory {
             children: Vec::new(),
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
         }
     }
 
@@ -165,7 +165,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform: AffineTransform::identity(),
             size: Self::DEFAULT_SIZE,
             corner_radius: RectangularCornerRadius::zero(),
@@ -188,7 +188,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: AffineTransform::identity(),
             fills: Paints::new([Self::default_solid_paint(Self::DEFAULT_COLOR)]),
@@ -208,7 +208,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: AffineTransform::identity(),
             size: Self::DEFAULT_SIZE,
@@ -229,7 +229,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: AffineTransform::identity(),
             size: Self::DEFAULT_SIZE,
@@ -251,7 +251,7 @@ impl NodeFactory {
             active: true,
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             effects: LayerEffects::default(),
             transform: AffineTransform::identity(),
             points: Vec::new(),
@@ -273,7 +273,7 @@ impl NodeFactory {
             opacity: Self::DEFAULT_OPACITY,
             blend_mode: LayerBlendMode::default(),
             effects: LayerEffects::default(),
-            mask_type: LayerMaskType::default(),
+            mask: None,
             transform: AffineTransform::identity(),
             size: Self::DEFAULT_SIZE,
             corner_radius: RectangularCornerRadius::zero(),
