+++ crates/grida-canvas/src/painter/mod.rs
@@ -5,5 +5,6 @@ pub mod gradient;
 pub mod image_filters;
 pub mod layer;
 pub mod paint;
+pub mod painter_debug_node;
 pub mod shadow;
 pub mod text_stroke;
