+++ crates/grida-canvas/src/painter/painter.rs
@@ -51,16 +51,11 @@ impl<'a> Painter<'a> {
         &self.path_cache
     }
 
-    /// Create a NodePainter that uses this Painter for its operations
-    pub fn node_painter(&self) -> NodePainter<'_> {
-        NodePainter::new(self)
-    }
-
     // ============================
     // === Helper Methods ========
     // ============================
 
-    fn with_transform_option<F: FnOnce()>(&self, transform: &Option<AffineTransform>, f: F) {
+    pub fn with_transform_option<F: FnOnce()>(&self, transform: &Option<AffineTransform>, f: F) {
         if let Some(transform) = transform {
             self.with_transform(&transform.matrix, f);
         } else {
@@ -69,7 +64,7 @@ impl<'a> Painter<'a> {
     }
 
     /// Save/restore transform state and apply a 2×3 matrix
-    fn with_transform<F: FnOnce()>(&self, transform: &[[f32; 3]; 2], f: F) {
+    pub fn with_transform<F: FnOnce()>(&self, transform: &[[f32; 3]; 2], f: F) {
         let canvas = self.canvas;
         canvas.save();
         canvas.concat(&sk::sk_matrix(*transform));
@@ -78,7 +73,7 @@ impl<'a> Painter<'a> {
     }
 
     /// If opacity < 1.0, wrap drawing in a save_layer_alpha; else draw directly.
-    fn with_opacity<F: FnOnce()>(&self, opacity: f32, f: F) {
+    pub fn with_opacity<F: FnOnce()>(&self, opacity: f32, f: F) {
         let canvas = self.canvas;
         if opacity < 1.0 {
             canvas.save_layer_alpha(None, (opacity * 255.0) as u32);
@@ -90,7 +85,7 @@ impl<'a> Painter<'a> {
     }
 
     /// If blend mode is not Normal, wrap drawing in a save_layer with blend mode; else draw directly.
-    fn with_blendmode<F: FnOnce()>(&self, layer_blend_mode: LayerBlendMode, f: F) {
+    pub fn with_blendmode<F: FnOnce()>(&self, layer_blend_mode: LayerBlendMode, f: F) {
         let canvas = self.canvas;
 
         // let mut paint = SkPaint::default();
@@ -112,7 +107,7 @@ impl<'a> Painter<'a> {
     }
 
     /// Helper method to apply clipping to a region with optional corner radius
-    fn with_clip<F: FnOnce()>(&self, shape: &PainterShape, f: F) {
+    pub fn with_clip<F: FnOnce()>(&self, shape: &PainterShape, f: F) {
         let canvas = self.canvas;
         canvas.save();
 
@@ -256,7 +251,7 @@ impl<'a> Painter<'a> {
         canvas.restore(); // pop the clip
     }
 
-    fn cached_path(&self, id: &NodeId, data: &str) -> Rc<Path> {
+    pub fn cached_path(&self, id: &NodeId, data: &str) -> Rc<Path> {
         self.path_cache.borrow_mut().get_or_create(id, data)
     }
 
@@ -290,7 +285,7 @@ impl<'a> Painter<'a> {
         )
     }
 
-    fn draw_fills(&self, shape: &PainterShape, fills: &[Paint]) {
+    pub fn draw_fills(&self, shape: &PainterShape, fills: &[Paint]) {
         if fills.is_empty() {
             return;
         }
@@ -303,7 +298,7 @@ impl<'a> Painter<'a> {
         }
     }
 
-    fn draw_strokes(
+    pub fn draw_strokes(
         &self,
         shape: &PainterShape,
         strokes: &[Paint],
@@ -342,7 +337,7 @@ impl<'a> Painter<'a> {
     }
 
     /// Draw a shape applying all layer effects in the correct order.
-    fn draw_shape_with_effects<F: Fn()>(
+    pub fn draw_shape_with_effects<F: Fn()>(
         &self,
         effects: &LayerEffects,
         shape: &PainterShape,
@@ -375,7 +370,7 @@ impl<'a> Painter<'a> {
         }
     }
 
-    fn draw_text_span(
+    pub fn draw_text_span(
         &self,
         id: &NodeId,
         text: &str,
@@ -432,7 +427,7 @@ impl<'a> Painter<'a> {
         );
     }
 
-    fn draw_text_paragraph(
+    pub fn draw_text_paragraph(
         &self,
         paragraph: &Rc<RefCell<textlayout::Paragraph>>,
         strokes: &[Paint],
@@ -663,455 +658,6 @@ impl<'a> Painter<'a> {
     pub fn draw_layer_list(&self, list: &LayerList) {
         for layer in &list.layers {
             self.draw_layer(layer);
-        }
-    }
-}
-
-/// A painter specifically for drawing nodes, using the main Painter for operations.
-/// This separates node-specific drawing logic from the main Painter while maintaining
-/// the ability to test golden outputs.
-pub struct NodePainter<'a> {
-    painter: &'a Painter<'a>,
-}
-
-impl<'a> NodePainter<'a> {
-    /// Create a new NodePainter that uses the given Painter
-    pub fn new(painter: &'a Painter<'a>) -> Self {
-        Self { painter }
-    }
-
-    /// Draw a RectangleNode, respecting its transform, effect, fill, stroke, blend mode, opacity
-    pub fn draw_rect_node(&self, node: &RectangleNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let shape = build_shape(&IntrinsicSizeNode::Rectangle(node.clone()));
-            self.painter
-                .draw_shape_with_effects(&node.effects, &shape, || {
-                    self.painter.with_opacity(node.opacity, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            self.painter.draw_fills(&shape, &node.fills);
-                            self.painter.draw_strokes(
-                                &shape,
-                                &node.strokes,
-                                node.stroke_width,
-                                node.stroke_align,
-                                node.stroke_dash_array.as_ref(),
-                            );
-                        });
-                    });
-                });
-        });
-    }
-
-    /// Draw an ImageNode, respecting transform, effect, rounded corners, blend mode, opacity
-    pub fn draw_image_node(&self, node: &ImageNodeRec) -> bool {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let shape = build_shape(&IntrinsicSizeNode::Image(node.clone()));
-
-            self.painter
-                .draw_shape_with_effects(&node.effects, &shape, || {
-                    self.painter.with_opacity(node.opacity, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            // Use the single image fill directly - aligns with web development patterns
-                            // where <img> elements have one image source
-                            // Create the Paint wrapper once and reuse the reference
-                            let image_paint = Paint::Image(node.fill.clone());
-                            self.painter
-                                .draw_fills(&shape, std::slice::from_ref(&image_paint));
-                            if !node.strokes.is_empty() {
-                                self.painter.draw_strokes(
-                                    &shape,
-                                    &node.strokes,
-                                    node.stroke_width,
-                                    node.stroke_align,
-                                    node.stroke_dash_array.as_ref(),
-                                );
-                            }
-                        });
-                    });
-                });
-        });
-        true
-    }
-
-    /// Draw an EllipseNode
-    pub fn draw_ellipse_node(&self, node: &EllipseNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let shape = build_shape(&IntrinsicSizeNode::Ellipse(node.clone()));
-            self.painter
-                .draw_shape_with_effects(&node.effects, &shape, || {
-                    self.painter.with_opacity(node.opacity, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            self.painter.draw_fills(&shape, &node.fills);
-                            self.painter.draw_strokes(
-                                &shape,
-                                &node.strokes,
-                                node.stroke_width,
-                                node.stroke_align,
-                                node.stroke_dash_array.as_ref(),
-                            );
-                        });
-                    });
-                });
-        });
-    }
-
-    /// Draw a LineNode
-    pub fn draw_line_node(&self, node: &LineNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let shape = build_shape(&IntrinsicSizeNode::Line(node.clone()));
-
-            self.painter.with_opacity(node.opacity, || {
-                self.painter.with_blendmode(node.blend_mode, || {
-                    self.painter.draw_strokes(
-                        &shape,
-                        &node.strokes,
-                        node.stroke_width,
-                        node.get_stroke_align(),
-                        node.stroke_dash_array.as_ref(),
-                    );
-                });
-            });
-        });
-    }
-
-    pub fn draw_vector_node(&self, node: &VectorNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let path = node.to_path();
-            let stroke_align = node.get_stroke_align();
-            let shape = PainterShape::from_path(path);
-            self.painter
-                .draw_shape_with_effects(&node.effects, &shape, || {
-                    self.painter.with_opacity(node.opacity, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            if !node.fills.is_empty() {
-                                self.painter.draw_fills(&shape, &node.fills);
-                            }
-                            self.painter.draw_strokes(
-                                &shape,
-                                &node.strokes,
-                                node.stroke_width,
-                                stroke_align,
-                                node.stroke_dash_array.as_ref(),
-                            );
-                        });
-                    });
-                });
-        });
-    }
-
-    /// Draw a PathNode (SVG path data)
-    pub fn draw_path_node(&self, node: &SVGPathNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let path = self.painter.cached_path(&node.id, &node.data);
-            let shape = PainterShape::from_path((*path).clone());
-            self.painter
-                .draw_shape_with_effects(&node.effects, &shape, || {
-                    self.painter.with_opacity(node.opacity, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            if !node.fills.is_empty() {
-                                self.painter.draw_fills(&shape, &node.fills);
-                            }
-                            if !node.strokes.is_empty() {
-                                self.painter.draw_strokes(
-                                    &shape,
-                                    &node.strokes,
-                                    node.stroke_width,
-                                    node.stroke_align,
-                                    node.stroke_dash_array.as_ref(),
-                                );
-                            }
-                        });
-                    });
-                });
-        });
-    }
-
-    /// Draw a PolygonNode (arbitrary polygon with optional corner radius)
-    pub fn draw_polygon_node(&self, node: &PolygonNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let path = node.to_path();
-            let shape = PainterShape::from_path(path.clone());
-            self.painter
-                .draw_shape_with_effects(&node.effects, &shape, || {
-                    self.painter.with_opacity(node.opacity, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            self.painter.draw_fills(&shape, &node.fills);
-                            self.painter.draw_strokes(
-                                &shape,
-                                &node.strokes,
-                                node.stroke_width,
-                                node.stroke_align,
-                                node.stroke_dash_array.as_ref(),
-                            );
-                        });
-                    });
-                });
-        });
-    }
-
-    /// Draw a RegularPolygonNode by converting to a PolygonNode
-    pub fn draw_regular_polygon_node(&self, node: &RegularPolygonNodeRec) {
-        let points = node.to_points();
-
-        let polygon = PolygonNodeRec {
-            id: node.id.clone(),
-            name: node.name.clone(),
-            active: node.active,
-            opacity: node.opacity,
-            blend_mode: node.blend_mode,
-            mask_type: node.mask_type,
-            transform: node.transform,
-            points,
-            corner_radius: node.corner_radius,
-            fills: node.fills.clone(),
-            strokes: node.strokes.clone(),
-            stroke_width: node.stroke_width,
-            stroke_align: node.stroke_align,
-            effects: node.effects.clone(),
-            stroke_dash_array: node.stroke_dash_array.clone(),
-        };
-
-        self.draw_polygon_node(&polygon);
-    }
-
-    /// Draw a RegularStarPolygonNode by converting to a PolygonNode
-    pub fn draw_regular_star_polygon_node(&self, node: &RegularStarPolygonNodeRec) {
-        let points = node.to_points();
-
-        let polygon = PolygonNodeRec {
-            id: node.id.clone(),
-            name: node.name.clone(),
-            active: node.active,
-            opacity: node.opacity,
-            blend_mode: node.blend_mode,
-            mask_type: node.mask_type,
-            transform: node.transform,
-            points,
-            corner_radius: node.corner_radius,
-            fills: node.fills.clone(),
-            strokes: node.strokes.clone(),
-            stroke_width: node.stroke_width,
-            stroke_align: node.stroke_align,
-            effects: node.effects.clone(),
-            stroke_dash_array: node.stroke_dash_array.clone(),
-        };
-
-        self.draw_polygon_node(&polygon);
-    }
-
-    /// Draw a TextSpanNode (simple text block)
-    pub fn draw_text_span_node(&self, node: &TextSpanNodeRec) {
-        if node.fills.is_empty() {
-            return;
-        }
-        self.painter.with_transform(&node.transform.matrix, || {
-            self.painter.with_opacity(node.opacity, || {
-                self.painter.with_blendmode(node.blend_mode, || {
-                    self.painter.draw_text_span(
-                        &node.id,
-                        &node.text,
-                        &node.width,
-                        &node.height,
-                        &node.max_lines,
-                        &node.ellipsis,
-                        &node.fills,
-                        &node.strokes,
-                        node.stroke_width,
-                        &node.stroke_align,
-                        &node.text_align,
-                        &node.text_align_vertical,
-                        &node.text_style,
-                    );
-                });
-            });
-        });
-    }
-
-    /// Draw a ContainerNode (background + stroke + children)
-    pub fn draw_container_node_recursively(
-        &self,
-        node: &ContainerNodeRec,
-        repository: &NodeRepository,
-        cache: &GeometryCache,
-    ) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            self.painter.with_opacity(node.opacity, || {
-                let shape = build_shape(&IntrinsicSizeNode::Container(node.clone()));
-
-                // Draw effects, fills, children (with optional clipping), then strokes last
-                self.painter
-                    .draw_shape_with_effects(&node.effects, &shape, || {
-                        self.painter.with_blendmode(node.blend_mode, || {
-                            // Paint fills first
-                            self.painter.draw_fills(&shape, &node.fills);
-
-                            // Children are drawn next; if `clip` is enabled we push
-                            // a clip region for the container's shape so that
-                            // descendants are clipped but the container's own stroke
-                            // remains unaffected.
-                            if node.clip {
-                                self.painter.with_clip(&shape, || {
-                                    for child_id in &node.children {
-                                        if let Some(child) = repository.get(child_id) {
-                                            self.draw_node_recursively(child, repository, cache);
-                                        }
-                                    }
-                                });
-                            } else {
-                                for child_id in &node.children {
-                                    if let Some(child) = repository.get(child_id) {
-                                        self.draw_node_recursively(child, repository, cache);
-                                    }
-                                }
-                            }
-
-                            // Finally paint the stroke so it is not clipped by the
-                            // container's own clip and always renders above children.
-                            self.painter.draw_strokes(
-                                &shape,
-                                &node.strokes,
-                                node.stroke_width,
-                                node.stroke_align,
-                                node.stroke_dash_array.as_ref(),
-                            );
-                        });
-                    });
-            });
-        });
-    }
-
-    pub fn draw_error_node(&self, node: &ErrorNodeRec) {
-        self.painter.with_transform(&node.transform.matrix, || {
-            let shape = build_shape(&IntrinsicSizeNode::Error(node.clone()));
-
-            // Create a red fill paint
-            let fill = Paint::Solid(SolidPaint {
-                color: CGColor(255, 0, 0, 51), // Semi-transparent red
-                blend_mode: BlendMode::Normal,
-            });
-            let stroke = Paint::Solid(SolidPaint {
-                color: CGColor(255, 0, 0, 255), // Solid red
-                blend_mode: BlendMode::Normal,
-            });
-
-            self.painter.with_opacity(node.opacity, || {
-                self.painter.draw_fills(&shape, std::slice::from_ref(&fill));
-                self.painter.draw_strokes(
-                    &shape,
-                    std::slice::from_ref(&stroke),
-                    1.0,
-                    StrokeAlign::Inside,
-                    None,
-                );
-            });
-        });
-    }
-
-    /// Draw a GroupNode: no shape of its own, only children, but apply transform + opacity
-    pub fn draw_group_node_recursively(
-        &self,
-        node: &GroupNodeRec,
-        repository: &NodeRepository,
-        cache: &GeometryCache,
-    ) {
-        self.painter.with_transform_option(&node.transform, || {
-            self.painter.with_opacity(node.opacity, || {
-                for child_id in &node.children {
-                    if let Some(child) = repository.get(child_id) {
-                        self.draw_node_recursively(child, repository, cache);
-                    }
-                }
-            });
-        });
-    }
-
-    pub fn draw_boolean_operation_node_recursively(
-        &self,
-        node: &BooleanPathOperationNodeRec,
-        repository: &NodeRepository,
-        cache: &GeometryCache,
-    ) {
-        self.painter.with_transform_option(&node.transform, || {
-            if let Some(shape) = boolean_operation_shape(node, repository, cache) {
-                self.painter
-                    .draw_shape_with_effects(&node.effects, &shape, || {
-                        self.painter.with_opacity(node.opacity, || {
-                            self.painter.with_blendmode(node.blend_mode, || {
-                                if !node.fills.is_empty() {
-                                    self.painter.draw_fills(&shape, &node.fills);
-                                }
-                                if !node.strokes.is_empty() {
-                                    self.painter.draw_strokes(
-                                        &shape,
-                                        &node.strokes,
-                                        node.stroke_width,
-                                        node.stroke_align,
-                                        node.stroke_dash_array.as_ref(),
-                                    );
-                                }
-                            });
-                        });
-                    });
-            } else {
-                for child_id in &node.children {
-                    if let Some(child) = repository.get(child_id) {
-                        self.draw_node_recursively(child, repository, cache);
-                    }
-                }
-            }
-        });
-    }
-
-    pub fn draw_node(&self, node: &LeafNode) {
-        if !node.active() {
-            return;
-        }
-        match node {
-            LeafNode::Error(n) => self.draw_error_node(n),
-            LeafNode::Rectangle(n) => self.draw_rect_node(n),
-            LeafNode::Ellipse(n) => self.draw_ellipse_node(n),
-            LeafNode::Polygon(n) => self.draw_polygon_node(n),
-            LeafNode::RegularPolygon(n) => self.draw_regular_polygon_node(n),
-            LeafNode::TextSpan(n) => self.draw_text_span_node(n),
-            LeafNode::Line(n) => self.draw_line_node(n),
-            LeafNode::Image(n) => {
-                self.draw_image_node(n);
-            }
-            LeafNode::Vector(n) => self.draw_vector_node(n),
-            LeafNode::SVGPath(n) => self.draw_path_node(n),
-            LeafNode::RegularStarPolygon(n) => self.draw_regular_star_polygon_node(n),
-        }
-    }
-
-    /// Dispatch to the correct node‐type draw method
-    pub fn draw_node_recursively(
-        &self,
-        node: &Node,
-        repository: &NodeRepository,
-        cache: &GeometryCache,
-    ) {
-        if !node.active() {
-            return;
-        }
-        match node {
-            Node::Error(n) => self.draw_error_node(n),
-            Node::Group(n) => self.draw_group_node_recursively(n, repository, cache),
-            Node::Container(n) => self.draw_container_node_recursively(n, repository, cache),
-            Node::Rectangle(n) => self.draw_rect_node(n),
-            Node::Ellipse(n) => self.draw_ellipse_node(n),
-            Node::Polygon(n) => self.draw_polygon_node(n),
-            Node::RegularPolygon(n) => self.draw_regular_polygon_node(n),
-            Node::TextSpan(n) => self.draw_text_span_node(n),
-            Node::Line(n) => self.draw_line_node(n),
-            Node::Image(n) => {
-                self.draw_image_node(n);
-            }
-            Node::Vector(n) => self.draw_vector_node(n),
-            Node::SVGPath(n) => self.draw_path_node(n),
-            Node::BooleanOperation(n) => {
-                self.draw_boolean_operation_node_recursively(n, repository, cache)
-            }
-            Node::RegularStarPolygon(n) => self.draw_regular_star_polygon_node(n),
         }
     }
 }
